[draft] private metadata migration v2
