+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/privacy/storage/migration/PrivateStorageMigrationTest.java
@@ -91,7 +91,7 @@ public class PrivateStorageMigrationTest {
     privateStateRootResolver = new PrivateStateRootResolver(privateStateStorage);
 
     migration =
-        new PrivateStorageMigration(
+        new PrivateStorageMigrationV1(
             privateStateStorage,
             blockchain,
             enclave,
